Додано кидання сніжки

+++ app/src/main/java/com/game/kolas/mygame/GameObject.java
@@ -14,11 +14,11 @@ public abstract class GameObject {
     protected int height;
     protected boolean isBonus;
 
-    public void setX(int  x)
+    public void setX(float  x)
     {
         this.x = x;
     }
-    public void setY(int y)
+    public void setY(float y)
     {
         this.y = y;
     }
